Share the Obsidian shim with a host that is not a test
The website runs the same importers in a browser, which has no vault and no
`obsidian` module either. What the tests already stand in for is most of what
it needs, so the half that behaves the same wherever a conversion runs moves
to web/obsidian: the YAML, the markdown, the file classes, the vault held in
memory, and Obsidian's additions to the DOM.

tests/shims keeps what only a test wants - a platform that is neither Obsidian
nor a browser, and stubs that throw so a conversion reaching for the app or
the network fails the test that runs it.

Sharing it is the point rather than the tidiness: it makes the recorded output
a check on what the website produces, not only on what the suite does.

A move, so the guards that were written for linkedom now ask an element what
it has rather than a constructor's prototype - linkedom hands every element
the plain HTMLElement prototype, so HTMLTableRowElement having `cells` says
nothing about a <tr> from that document, while a browser's does. 847 tests
unchanged.
